Create CSVtoDataFrame_HTML_visualisation
The code compares the two csv files using pandas, by converting them into Data Frame. Using or assuming an unique column to be the primary column, the mismatched columns and values are listed out. The author is also working on visualizing the matched and unmatched records of the files in green and red respectively. However partial output could be shown currently.